+++ backend/src/main/java/com/sol/snappick/util/minio/ImageUploadRes.java
@@ -1,4 +1,4 @@
-package com.sol.snappick.util;
+package com.sol.snappick.util.minio;
 
 import lombok.Builder;
 import lombok.Getter;
